# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CATEDRALX - Chalamandra Labs SNC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /system/affective:
    post:
      operationId: recordAffective
      tags: [vitals]
      summary: Record an affective state event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffectiveBody"
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /signals:
    get:
      operationId: getSignals
      tags: [signals]
      summary: Get trading signals
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Signals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalsResponse"
...
  /invoke-strategy/{id}:
    post:
      operationId: invokeStrategy
      tags: [signals]
      summary: Invoke a strategy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy invoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyInvokeResponse"
...
components:
  schemas:
...
    AffectiveBody:
      type: object
      properties:
        source:
          type: string
      required: [source]
    Signal:
      type: object
      properties:
        id:
          type: integer
        strategy:
          type: string
        type:
          type: string
        message:
          type: string
        price:
          type: number
        confidence:
          type: integer
        created_at:
          type: string
      required: [strategy, type, message, price, confidence]
    SignalsResponse:
      type: object
      properties:
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        limited:
          type: boolean
        tier:
          type: string
      required: [signals, limited, tier]
...
    StrategyInvokeResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        id:
          type: integer
      required: [status, message, id]
...