# lib/api-zod/src/generated/types/strategyInvokeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CATEDRALX - Chalamandra Labs SNC API
 * OpenAPI spec version: 0.1.0
 */

export interface StrategyInvokeResponse {
  status: string;
  message: string;
  id: number;
}